==> analyses/survival/01-prepare-survival.Rmd <==
---
title: "Prepare splicing data for survival analyses"
output: 
  html_notebook:
    toc: TRUE
    toc_float: TRUE
author: Ryan Corbett, Jo Lynne Rokita
date: 2023
params:
  plot_ci: TRUE
---

**Purpose:** 

- Add sample survival data to splicing index file to run survival analyses
- Add additional subtype data from methylation to improve subtype-specific analyses

## Setup

#### Packages and functions

load libraries and set directories
```{r Set up library}
library(survival)
library(ggpubr)
library(tidyverse)

# Set directories
root_dir <- rprojroot::find_root(rprojroot::has_dir(".git"))
data_dir <- file.path(root_dir, "data")
analysis_dir <- file.path(root_dir, "analyses", "survival")
input_dir <- file.path(analysis_dir, "input")
results_dir <- file.path(analysis_dir, "results")

# Magrittr pipe
`%>%` <- dplyr::`%>%`
```

Set file paths

```{r}
sbi_A3SS_file <- file.path(root_dir, "analyses", "splicing_index", "results", "splicing_index.A3SS.txt")
sbi_A5SS_file <- file.path(root_dir, "analyses", "splicing_index", "results", "splicing_index.A5SS.txt")
sbi_RI_file <- file.path(root_dir, "analyses", "splicing_index", "results", "splicing_index.RI.txt")
sbi_SE_file <- file.path(root_dir, "analyses", "splicing_index", "results", "splicing_index.SE.txt")

histologies_file <- file.path(data_dir, "histologies-plot-group.tsv")

```

Read in splicing index file, histologies, and updated survival files
```{r}
si_A3SS <- read_tsv(sbi_A3SS_file) %>%
  rename("Kids_First_Biospecimen_ID"= Sample,
         SI_A3SS = SI)
si_A5SS <- read_tsv(sbi_A5SS_file) %>%
  dplyr::filter(!is.na(Histology)) %>%
  rename("Kids_First_Biospecimen_ID"= Sample,
         SI_A5SS = SI) %>%
  dplyr::mutate(SI_A5SS = as.double(SI_A5SS))
si_RI <- read_tsv(sbi_RI_file) %>%
  dplyr::filter(!is.na(Histology)) %>%
  rename("Kids_First_Biospecimen_ID"= Sample,
         SI_RI = SI) %>%
  dplyr::mutate(SI_RI = as.double(SI_RI))
si_SE <- read_tsv(sbi_SE_file) %>%
  rename("Kids_First_Biospecimen_ID"= Sample,
         SI_SE = SI)

splice_index <- si_A3SS %>%
  left_join(si_A5SS[,c("Kids_First_Biospecimen_ID", "SI_A5SS")]) %>%
  left_join(si_RI[,c("Kids_First_Biospecimen_ID", "SI_RI")]) %>%
  left_join(si_SE[,c("Kids_First_Biospecimen_ID", "SI_SE")])

```

Use `diagnoses` to code in `EFS_status` (event-free status) column indicating event/no event

```{r}
hist <- read_tsv(histologies_file) %>%
  filter(experimental_strategy == "RNA-Seq")

# Add histology columns to splice index, and code EFS_status using following logic
splice_index <- splice_index %>%
  filter(Kids_First_Biospecimen_ID %in% hist$Kids_First_Biospecimen_ID) %>%
  left_join(hist[,c("Kids_First_Biospecimen_ID", "Kids_First_Participant_ID", 
                    "molecular_subtype",
                    "extent_of_tumor_resection", 
                    "EFS_days", "EFS_event_type", 
                    "OS_days", "OS_status")]) %>%
  distinct(Kids_First_Participant_ID, .keep_all = T) %>%
  dplyr::mutate(EFS_days = as.numeric(EFS_days),
         OS_days = as.numeric(OS_days)) %>%
  dplyr::mutate(OS_years = OS_days/365.25,
         EFS_years = EFS_days/365.25) %>%
  
  # code EFS status using following logic
  dplyr::mutate(EFS_status = case_when(
    EFS_event_type %in% c("Not Applicable", "Not Reported") ~ "NO EVENT",
    !is.na(EFS_event_type) ~ "EVENT",
    TRUE ~ NA_character_
  ))

specimens <- read_tsv(file.path(data_dir, "independent-specimens.rnaseqpanel.primary.tsv"))

dup_pts <- splice_index[duplicated(splice_index$Kids_First_Participant_ID),]$Kids_First_Participant_ID

splice_index <- splice_index %>%
  dplyr::filter(Kids_First_Biospecimen_ID %in% specimens$Kids_First_Biospecimen_ID | !Kids_First_Participant_ID %in% dup_pts)
```

Define column `mol_sub_group` that will consolidate LGG and GNG subtypes

```{r}

splice_index <- splice_index %>%
  filter(!grepl("To be classified", molecular_subtype)) %>%
  mutate(mol_sub_group = case_when(
    Histology %in% c("Low-grade glioma", "Mixed neuronal-glial tumor")  & grepl("SEGA", molecular_subtype) ~ "SEGA",
    Histology %in% c("Low-grade glioma", "Mixed neuronal-glial tumor") & grepl("wildtype", molecular_subtype) ~ "Wildtype",
    Histology %in% c("Low-grade glioma", "Mixed neuronal-glial tumor") & grepl("V600E", molecular_subtype) ~ "BRAF V600E",
    Histology %in% c("Low-grade glioma", "Mixed neuronal-glial tumor") & grepl("KIAA1549-BRAF", molecular_subtype) ~ "BRAF fusion",
    Histology %in% c("Low-grade glioma", "Mixed neuronal-glial tumor") ~ "Other alteration",
    !is.na(molecular_subtype) ~ molecular_subtype
  ))

```

Create subtype df consisting of histology group molecular subtypes for which survival will be assessed

```{r}

groups <- c("Atypical Teratoid Rhabdoid Tumor", "Craniopharyngioma", "Ependymoma", "Mixed neuronal-glial tumor", "DIPG or DMG", "Other high-grade glioma", "Low-grade glioma", "Medulloblastoma")

# folder names
folder <- c("ATRT", "CPG", "EPN", "GNG", "DMG", "HGG", "LGG", "MB")
names(folder) <- groups
  
# create names vector for output files
plot_names <- c("atrt", "cpg", "epn", "gng", "dmg", "hgg", "lgg", "mb")
names(plot_names) <- groups

# define subtype df with histology, subtype, and file names
subtype_df <- splice_index %>%
  mutate(subtype = gsub("^HGG, |^ATRT, |CRANIO, |DHG, H3 |EPN, |IHG, |MB, |, TP53|DMG, H3 | alteration", "", mol_sub_group)) %>%
  group_by(Histology, subtype) %>%
  filter(!is.na(subtype),
         Histology %in% groups,
         !is.na(OS_days) | !is.na(EFS_days)) %>%
  count() %>%
  filter(n >=16) %>%
  dplyr::rename(hist = Histology) %>%
  left_join(tibble(hist = groups, names = plot_names), by = "hist") %>%
  mutate(names = coalesce(names, NA_character_)) %>%
  left_join(tibble(hist = groups, foldername = folder), by = "hist") %>%
  dplyr::mutate(subtype_name = str_replace_all(subtype, " ", "-"))  %>%
  write_tsv(file.path(analysis_dir, "results", "subtypes-for-survival.tsv"))

```

Save to output
```{r}
# Save output
write_tsv(splice_index,
          file.path(results_dir, "splicing_indices_with_survival.tsv"))
```
